\name{stlouis}
\alias{stlouis}
\docType{data}
\title{Precinct-Level Data from 1993 St. Louis Mayoral Election}
\description{
Precinct-Level Data from 1993 St. Louis Mayoral Election
}
\usage{data(stlouis)}
\format{
  A data frame with 349 observations on the following 7 variables. Each observation is a precinct.
  \describe{
    \item{\code{bvap}}{Black voting age population}
    \item{\code{ovap}}{Non-black (other) voting age population}
    \item{\code{Bosley}}{Votes for Bosley}
    \item{\code{Roberts}}{Votes for Roberts}
    \item{\code{Ribaudo}}{Votes for Ribaudo}
    \item{\code{Villa}}{Votes for Villa}
    \item{\code{NoVote}}{Number of non-voters}
  }
}

\source{
  Personal communication with Brady Baybeck.
}

\keyword{datasets}
